requirements.txt
tensorflow-cpu
numpy>=1.19.5
Keras>=2.4.3
opencv-contrib-python-headless>=4.5.1.48
Flask>=1.1.2
gunicorn>=20.1.0


